requirements.txt
gradio
requests
python-dotenv
pandas
langchain
openai

